feat(exercise): corrigindo busca banco de dados
@patriciatersariol @estherwl

46
