+++ src/controllers/exercise/index.ts
@@ -7,12 +7,12 @@ import { Exercise } from "@models/entity/Exercise"
 import { getRepository } from "typeorm"
 import { importSpreadSheet } from "@service/google-spreadsheet"
 import { response } from 'express'
-//import { ExerciseService } from "@service/exercise/ExerciseService"
+import { ExerciseService } from "@service/exercise/ExerciseService"
 
 
 const evaluationService = new EvaluationService()
 const httpResponseHandler = new HttpResponseHandler()
-//const exerciseService = new ExerciseService()
+const exerciseService = new ExerciseService()
 
 export const createEvaluation = async (request, response) => {
   try {
@@ -100,25 +100,15 @@ export const importExercises = async (request, response) => {
   
 }
 
-// export const getAllExercises = async (request, response) => {
-//   try {
-//     const {page, count} = request.query
-//     const id = request.params.id
-//     const exercises = await exerciseService.getAllExercisesService(page, count, id)
-//     return httpResponseHandler.createSuccessResponse(message.FOUND, exercises, response)
-//   }
-//   catch (error) {
-//     return httpResponseHandler.createErrorResponse(error, response)
-//   }
-// }
 export const getExerciseByHiringProcessId = async (req, res) => {
-  const { hiringProcessId } = req.query
-  const exerciseRepository = getRepository(Exercise)
-  const result = await exerciseRepository.createQueryBuilder()
-    .select("exercise")
-    .from(Exercise, "exercise")
-    .leftJoinAndSelect("exercise.hiringProcess", "hiringProcess")
-    .getMany()
-  // .where("exercise.hiring_process_id = :id", { id: hiringProcessId })
-  return res.json({ hiringProcessId, result })
+  const { hiringProcessId } = req.params
+  const {page, count} = req.query
+  try {
+    const result = await exerciseService.getAllExercisesService(page, count, hiringProcessId)
+    return httpResponseHandler.createSuccessResponse(message.SUCCESS, { hiringProcessId, result }, response)
+    //return res.json({ hiringProcessId, result })
+  } 
+  catch (error) {
+    return httpResponseHandler.createErrorResponse(error, response)
+  }
 }
