+++ src/routes.ts
@@ -28,5 +28,5 @@ export const defineRoutes = (app) => {
   app.post('/exercise', createEvaluation)
   app.patch('/exercise/:id', editEvaluation)
   app.post('/exercise/hiring_process/:id', importExercises)
-  app.get('/exercise', getExerciseByHiringProcessId)
+  app.get('/exercise/:hiringProcessId', getExerciseByHiringProcessId)
 }
